delete

+++ src/main/java/com/example/lab2vscode/service/RegionService.java
@@ -5,7 +5,9 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.lab2vscode.model.Country;
 import com.example.lab2vscode.model.Region;
+import com.example.lab2vscode.repository.CountryRepository;
 import com.example.lab2vscode.repository.RegionRepository;
 
 import lombok.AllArgsConstructor;
@@ -14,6 +16,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class RegionService {
     private RegionRepository regionRepository;
+    private CountryRepository countryRepository;
 
     public Region createRegion(Region regionModel) {
         return regionRepository.save(regionModel);
@@ -32,7 +35,17 @@ public class RegionService {
     }
 
     public void deleteRegion(Integer regionId) {
-        regionRepository.deleteById(regionId);
+        Optional<Region> region = regionRepository.findById(regionId);
+        if(region.isPresent())
+        {
+            List<Country> countries = region.get().getCountries().stream().toList();
+            for(Country country : countries)
+            {
+                country.setRegion(null);
+                countryRepository.save(country);
+            }
+            regionRepository.deleteById(regionId);
+        }
     }
 
     public Region updateRegion(Integer regionId , Region regionDetails) {
